refactor: 레거시 코드 제거하고 어노테이션 기반으로 변경
