--- app/src/main/java/com/techcourse/ControllerHandlerAdapter.java
@@ -1,24 +1,0 @@
-package com.techcourse;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-
-public class ControllerHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(final Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response,
-                               final Object handler)
-            throws Exception {
-        String viewName = ((Controller) handler).execute(request, response);
-        return new ModelAndView(new JspView(viewName));
-    }
-}
