+++ app/src/main/java/com/techcourse/controller/IndexController.java
@@ -1,0 +1,19 @@
+package com.techcourse.controller;
+
+import static nextstep.web.support.RequestMethod.GET;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.view.JspView;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+
+@Controller
+public class IndexController {
+
+    @RequestMapping(value = "/", method = GET)
+    public ModelAndView index(final HttpServletRequest req, final HttpServletResponse res) {
+        return new ModelAndView(new JspView("/index.jsp"));
+    }
+}
