+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -1,15 +1,22 @@
 package com.techcourse.controller;
 
+import static nextstep.web.support.RequestMethod.GET;
+import static nextstep.web.support.RequestMethod.POST;
+
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import com.techcourse.repository.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.mvc.view.JspView;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class LoginController implements Controller {
+@Controller
+public class LoginController {
 
     private static final Logger log = LoggerFactory.getLogger(LoginController.class);
 
@@ -23,27 +30,36 @@ public class LoginController implements Controller {
         this.userRepository = userRepository;
     }
 
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        if (UserSession.isLoggedIn(req.getSession())) {
-            return "redirect:/index.jsp";
-        }
+    @RequestMapping(value = "/login", method = GET)
+    public ModelAndView showLoginPage(final HttpServletRequest req, final HttpServletResponse res) {
+        return UserSession.getUserFrom(req.getSession())
+                .map(user -> {
+                    log.info("logged in {}", user.getAccount());
+                    return new ModelAndView(new JspView("redirect:/index.jsp"));
+                })
+                .orElse(new ModelAndView(new JspView("/login.jsp")));
+    }
 
+    @RequestMapping(value = "/login", method = POST)
+    public ModelAndView login(final HttpServletRequest req, final HttpServletResponse res) {
+        if (UserSession.isLoggedIn(req.getSession())) {
+            return new ModelAndView(new JspView("redirect:/index.jsp"));
+        }
         return userRepository.findByAccount(req.getParameter("account"))
                 .map(user -> {
                     log.info("User : {}", user);
                     return login(req, user);
                 })
-                .orElse("redirect:/401.jsp");
+                .orElse(new ModelAndView(new JspView("redirect:/401.jsp")));
     }
 
-    private String login(final HttpServletRequest request, final User user) {
+    private ModelAndView login(final HttpServletRequest request, final User user) {
         if (user.checkPassword(request.getParameter("password"))) {
             final var session = request.getSession();
             session.setAttribute(UserSession.SESSION_KEY, user);
-            return "redirect:/index.jsp";
+            return new ModelAndView(new JspView("redirect:/index.jsp"));
         } else {
-            return "redirect:/401.jsp";
+            return new ModelAndView(new JspView("redirect:/401.jsp"));
         }
     }
 }
