+++ app/src/main/java/com/techcourse/controller/LogoutController.java
@@ -1,15 +1,22 @@
 package com.techcourse.controller;
 
+import static nextstep.web.support.RequestMethod.GET;
+
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
+import jakarta.servlet.http.HttpSession;
+import nextstep.mvc.view.JspView;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
 
-public class LogoutController implements Controller {
+@Controller
+public class LogoutController {
 
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        final var session = req.getSession();
+    @RequestMapping(value = "/logout", method = GET)
+    public ModelAndView logout(final HttpServletRequest req, final HttpServletResponse res) {
+        final HttpSession session = req.getSession();
         session.removeAttribute(UserSession.SESSION_KEY);
-        return "redirect:/";
+        return new ModelAndView(new JspView("redirect:/"));
     }
 }
